// chinesetexts/src/app/model/text.ts
export interface Text {
    id?: number;
    titleEnglish: string;
    titleSpanish: string;
    text: string;
    spanishTranslation: string;
    englishTranslation: string;
    level: string;
    englishDescription: string;
    spanishDescription: string;
    creationDate: string;
    translation: string;
}